Completely rebuilt system of task list and progress updates; no more instances of hanged cells, dead views, or a list that gets stuck on a task. Also, new modeling system of task media, comments, and revisions is working perfectly. Just need to continue to remember that *realm objects always must be accessed from the correct thread!!*

+++ CommSync/CommSync/Models/CSTaskRevisionRealmModel.m
@@ -156,7 +156,7 @@
                    arc4random_uniform(25)+65,
                    arc4random_uniform(25)+65];
     
-    _revisionID = [NSString stringWithFormat:@"%@_%ld", U, sourceTask.revisions.count];
+    _revisionID = [NSString stringWithFormat:@"%@_%ld", U, (unsigned long)sourceTask.revisions.count];
     _revisionDate = [NSDate new];
 }
